cli: Import overrides from crucible-syntax into the test suite
