+++ crucible-cli/crucible-cli.cabal
@@ -149,7 +149,9 @@ test-suite crucible-cli-tests
     directory,
     filepath,
     lens,
+    mtl,
     parameterized-utils >= 0.1.7,
+    prettyprinter >= 1.7.0,
     tasty,
     tasty-hunit,
     tasty-golden,
